getWedding endpoint to work
